+++ src/main/java/shopping/auth/AuthArgumentResolver.java
@@ -2,11 +2,13 @@ package shopping.auth;
 
 import javax.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
 
+@Component
 public class AuthArgumentResolver implements HandlerMethodArgumentResolver {
 
     @Override
