+++ src/main/java/shopping/auth/JwtInterceptor.java
@@ -1,11 +1,9 @@
-package shopping.interceptor;
+package shopping.auth;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
-import shopping.auth.BearerExtractor;
-import shopping.auth.JwtHelper;
 import shopping.exception.ShoppingAuthenticationException;
 
 @Component
