+++ src/main/java/shopping/config/WebConfig.java
@@ -6,15 +6,18 @@ import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import shopping.auth.AuthArgumentResolver;
-import shopping.interceptor.JwtInterceptor;
+import shopping.auth.JwtInterceptor;
 
 @Configuration
 public class WebConfig implements WebMvcConfigurer {
 
     private final JwtInterceptor jwtInterceptor;
+    private final AuthArgumentResolver authArgumentResolver;
 
-    public WebConfig(final JwtInterceptor jwtInterceptor) {
+    public WebConfig(final JwtInterceptor jwtInterceptor,
+        final AuthArgumentResolver authArgumentResolver) {
         this.jwtInterceptor = jwtInterceptor;
+        this.authArgumentResolver = authArgumentResolver;
     }
 
     @Override
@@ -25,6 +28,6 @@ public class WebConfig implements WebMvcConfigurer {
 
     @Override
     public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new AuthArgumentResolver());
+        resolvers.add(authArgumentResolver);
     }
 }
